Fix issues in round1
