+++ .fossa.yml
@@ -894,7 +894,7 @@ targets:
       target: ':instrumentation:spring:spring-kafka-2.7:library'
     - type: gradle
       path: ./
-      target: ':instrumentation:spring:spring-pulsar-1.0:javaagent'
+      target: ':instrumentation:spring:spring-pulsar-1.2:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:spring:spring-rabbit-1.0:javaagent'
