+++ instrumentation/spring/spring-pulsar-1.2/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/spring/pulsar/v1_2/SpringPulsarTest.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.spring.pulsar.v1_2;
+
+import static io.opentelemetry.api.trace.SpanKind.CONSUMER;
+import static io.opentelemetry.api.trace.SpanKind.INTERNAL;
+import static io.opentelemetry.api.trace.SpanKind.PRODUCER;
+import static io.opentelemetry.instrumentation.testing.util.TelemetryDataUtil.orderByRootSpanKind;
+
+import io.opentelemetry.instrumentation.spring.pulsar.v1_0.AbstractSpringPulsarTest;
+
+class SpringPulsarTest extends AbstractSpringPulsarTest {
+
+  @Override
+  protected void assertSpringPulsar() {
+    testing.waitAndAssertSortedTraces(
+        orderByRootSpanKind(INTERNAL, CONSUMER),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span -> span.hasName("parent").hasNoParent(),
+                span ->
+                    span.hasName(OTEL_TOPIC + " publish")
+                        .hasKind(PRODUCER)
+                        .hasParent(trace.getSpan(0))
+                        .hasAttributesSatisfyingExactly(publishAttributes()),
+                span ->
+                    span.hasName(String.format("%s process", OTEL_TOPIC))
+                        .hasKind(CONSUMER)
+                        .hasParent(trace.getSpan(1))
+                        .hasAttributesSatisfyingExactly(processAttributes()),
+                span -> span.hasName("consumer").hasParent(trace.getSpan(2))),
+        trace ->
+            trace.hasSpansSatisfyingExactly(
+                span ->
+                    span.hasName(String.format("%s receive", OTEL_TOPIC))
+                        .hasKind(CONSUMER)
+                        .hasNoParent()
+                        .hasAttributesSatisfyingExactly(receiveAttributes())));
+  }
+}
